% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\docType{package}
\name{EurostatBDF}
\alias{EurostatBDF}
\alias{EurostatBDF-package}
\title{A package for downloading datasets from the Eurostat Bulk Download Facility.}
\description{
The workhorse function of the package is \code{getEurostat()}. This function returns a
list containing a Eurostat dataset, the code lists for the dataset, and the dataset metadata.
It is optional for the function to also return the raw data (as it was downloaded)
and the data in CSV form (containing codes and labels). The code behind this package also
serves as a backend to an EViews add-in created by the author.
}
\section{EurostatBDF functions}{

\code{getEurostat()}
}

