% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{getCodeLists}
\alias{getCodeLists}
\title{Get the code lists for a dataset}
\usage{
getCodeLists(tidy_data)
}
\arguments{
\item{tidy_data}{A tidy Eurostat dataset}
}
\description{
Creates a list containing all of the dataset dimensions including values and labels
}
\details{
The dataset dimensions are matched to the values in the corresponding dictionary files
}
\author{
Graeme Walsh
}
