% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{getColhead}
\alias{getColhead}
\title{Get column header properties (for use in EViews)}
\usage{
getColhead(csv_data)
}
\arguments{
\item{csv_data}{A tidy Eurostat dataset prepared to be exported as CSV file for EViews}
}
\description{
Gets the names and number of column headers (for use in EViews)
}
\details{
Returns the names and number of column headers for importing a CSV file into EViews
}
\author{
Graeme Walsh
}
