% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{getDic}
\alias{getDic}
\title{Get a dictionary file from the Eurostat Bulk Download Facility}
\usage{
getDic(dic_name)
}
\arguments{
\item{dic_name}{The name of the dictionary file of interest}
}
\description{
Reads a Eurostat dictionary file into R as a two-column dataframe
}
\details{
The dictionary file is read from the /dic/en sub-folder
}
\author{
Graeme Walsh
}
