% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{getEurostat}
\alias{getEurostat}
\title{Get a complete dataset from the Eurostat Bulk Download Facility}
\usage{
getEurostat(dataset_name, append_raw = FALSE, append_csv = FALSE)
}
\arguments{
\item{dataset_name}{The dataset code of interest}

\item{append_raw}{TRUE or FALSE}

\item{append_csv}{TRUE or FALSE}
}
\description{
Download data from Eurostat (including code lists and metadata)
}
\details{
The dataset is returned as a list containing data, code lists, and metadata. The raw data can also be returned.
}
\author{
Graeme Walsh
}
