% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{getEurostatRaw}
\alias{getEurostatRaw}
\title{Get a dataset in raw format from the Eurostat Bulk Download Facility}
\usage{
getEurostatRaw(dataset_name, format = "tsv")
}
\arguments{
\item{dataset_name}{The dataset code of interest}

\item{format}{The format of the dataset}
}
\description{
Download data from Eurostat
}
\details{
The dataset is downloaded from the Bulk Download Facility
}
\author{
Graeme Walsh
}
