% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{getMeta}
\alias{getMeta}
\title{Get the metadata associated with a dataset}
\usage{
getMeta(dataset_name)
}
\arguments{
\item{dataset_name}{The dataset code of interest}
}
\description{
Reads the metadata for a dataset from the table of contents file
}
\details{
The metdata contains information such as title, code, last update, etc.
}
\author{
Graeme Walsh
}
