% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{getPeriod}
\alias{getPeriod}
\title{Get the time period and frequency of a dataset}
\usage{
getPeriod(dataset_name)
}
\arguments{
\item{dataset_name}{The dataset code of interest}
}
\description{
Gets the time period and frequency information from the table of contents file
}
\details{
Returns the frequency, start date, and end date of the dataset
}
\author{
Graeme Walsh
}
