% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{getTdic}
\alias{getTdic}
\title{Get the table dictionary file from the Eurostat Bulk Download Facility}
\usage{
getTdic()
}
\description{
Reads the Eurostat table dictionary file into R as a two-column dataframe
}
\details{
The table dictionary file is read from the /dic/en sub-folder
}
\author{
Graeme Walsh
}
