% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{getToc}
\alias{getToc}
\title{Get the table of contents file from the Eurostat Bulk Download Facility}
\usage{
getToc()
}
\description{
Reads the Eurostat table of contents file into R as a eight-column dataframe
}
\details{
The table of contents file is read from the home directory
}
\author{
Graeme Walsh
}
