% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{prepareCSV}
\alias{prepareCSV}
\title{Prepare a dataset to be saved in CSV format (add codes and labels to dataframe)}
\usage{
prepareCSV(tidy_data)
}
\arguments{
\item{tidy_data}{A tidy Eurostat dataset}
}
\description{
Prepares a dataset to be saved in CSV format (containing codes and labels)
}
\details{
Codes and labels are added to the dataframe
}
\author{
Graeme Walsh
}
