% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EurostatBDF.R
\name{tidyRawData}
\alias{tidyRawData}
\title{Tidy a raw format Eurostat dataset (remove flags, etc.)}
\usage{
tidyRawData(raw_data)
}
\arguments{
\item{raw_data}{The raw dataset of interest}
}
\description{
Tidy a raw format Eurostat dataset (remove flags, etc.)
}
\details{
Flags are removed, missing values are replaced (: to NA), and columns are converted from descending to ascending
}
\author{
Graeme Walsh
}
